Remove Win32 preview state types from skin select menu

+++ Minecraft.Client/Platform/Common/UI/UIScene_SkinSelectMenu.h
@@ -11,7 +11,7 @@ private:
 	static const WCHAR *wchDefaultNamesA[eDefaultSkins_Count];
 
 	// 4J Stu - How many to show on each side of the main control
-	static const BYTE sidePreviewControls = 4;
+	static const int sidePreviewControls = 4;
 
 #ifdef __PSVITA__
 	enum ETouchInput
